refactor: corrected file name of codebuild deploy
